rust: drm/kms: Add RawConnector and RawConnectorState
Now that we have more then one way to refer to connectors, we also want to
ensure that any methods which are common to any kind of connector type can
be used on all connector representations. This is where RawConnector and
RawConnectorState come in: we implement these traits for any type which
implements AsRawConnector or AsRawConnectorState respectively.

Signed-off-by: Lyude Paul <lyude@redhat.com>
(cherry picked from commit 58d97f5488369ac2dab3cf9271603bc9c71c1f98)