rust: drm/kms: Add RawPlaneState::atomic_helper_check()
Add a binding for drm_atomic_helper_check_plane_state(). Since we want to
make sure that the user is passing in the new state for a Crtc instead of
an old state, we explicitly ask for a reference to a CrtcStateMutator.

Signed-off-by: Lyude Paul <lyude@redhat.com>
(cherry picked from commit 5ea2c173ab6bf2d7032aa97e8284a892eaeeead1)