rust: drm/kms: Introduce impl_atomic_state_token_ops!
In order to add callbacks for various KMS hooks such as
drm_crtc_helper_funcs, we want to allow for implementers to be able to
acquire the current old and new object atomic states within a given
callback and do so unconditionally since we can assume ahead of time that
they'll always be present in the global atomic state. While we could pass
the old and new state directly to the functions, the C side of DRM already
did this in the past and chose to remove these arguments for the simple
reason that you don't always need access to these states in a callback.

This means we need a way to be able to lazily get the old and new state
without introducing a conditional, which brings us to the
impl_atomic_state_token_ops macro. Using Crtc and atomic_check as an
example, we can use impl_atomic_state_token_ops to introduce a convenient
CrtcAtomicCheck token type. This type acts as a promise that the old and
new atomic state for the respective Crtc are in-fact present in the global
atomic state, even though they haven't been fetched yet.

This token provides access to the crtc in question, and can be consumed to
access one or multiple of:

  * The global atomic state (AtomicStateComposer, AtomicStateMutator, etc.)
  * The old atomic state
  * The new atomic state

like so:

  impl DriverCrtc for CoolCrtc {
    // …
    fn atomic_check(check: CrtcAtomicCheck<'_>) -> Result {
      let crtc = check.crtc;
      let (state, old_state, new_state) = check.take_all();
      // Or
      // let state = check.take_atomic_state();
      // etc.
      // …
    }
  }

So, introduce the impl_atomic_state_token_ops macro - which handles
defining this type since doing so is rather redundant for each KMS object.

Signed-off-by: Lyude Paul <lyude@redhat.com>
(cherry picked from commit 4d28e81f6d3489d971dc1ec6fb6511e4f39a40a5)