rust: drm/kms: Add UnregisteredConnector::attach_encoder()
This adds a simple binding for completing the last step of creating a DRM
connector - attaching its encoder. This function should only be called
before the connector is registered, and DRM should enforce this itself by
returning an error if a driver tries to add an encoder to an
already-registered DRM connector.

Note that unlike most of the methods we'll be adding to DRM mode objects,
this is directly implemented on the Connector<T> type since I don't really
think it would make sense for us to allow this operation on an
OpaqueConnector (a DRM connector without a known DriverConnector
implementation, something we'll be adding in the next few commits).

Signed-off-by: Lyude Paul <lyude@redhat.com>

---

V3:
* Move to UnregisteredConnector interface
* Improve safety comments

(cherry picked from commit 322a9b8d699b008a8d37d4ad273f5b3fc14f4c15)